move y type casting to the end

+++ rastervision_pytorch_learner/rastervision/pytorch_learner/dataset/transform.py
@@ -23,10 +23,11 @@ def classification_transformer(inp: Tuple[Any, Any],
                                ) -> Tuple[np.ndarray, np.ndarray]:
     """Apply transform to image only."""
     x, y = inp
-    x, y = np.array(x), np.array(y, dtype=np.long)
+    x, y = np.array(x), np.array(y)
     if transform is not None:
         out = transform(image=x)
         x = out['image']
+    y = y.astype(np.long)
     return x, y
 
 
@@ -166,10 +167,11 @@ def semantic_segmentation_transformer(inp: Tuple[Any, Any],
                                       ) -> Tuple[np.ndarray, np.ndarray]:
     """Apply transform to image and mask."""
     x, y = inp
-    x, y = np.array(x), np.array(y, dtype=np.long)
+    x, y = np.array(x), np.array(y)
     if transform is not None:
         out = transform(image=x, mask=y)
         x, y = out['image'], out['mask']
+    y = y.astype(np.long)
     return x, y
